- implementation des liaisons users / groups / projects a double sens
Reste a implementer les test unitaires et les select / deselect all js a regler

refs #148, #149

+++ www/USVN/modules/admin/controllers/GroupController.php
@@ -24,7 +24,8 @@ class admin_GroupController extends admin_IndexController
 {
 	public static function getGroupDataFromPost()
 	{
-		$group = array('groups_name'	=> $_POST['groups_name']);
+		$group = array('groups_name'			=> $_POST['groups_name'],
+						'groups_description'	=> $_POST['groups_description']);
 		if (isset($_POST['groups_id']) && !empty($_POST['groups_id'])) {
 			$group['groups_id'] = $_POST['groups_id'];
 		}
@@ -40,7 +41,8 @@ class admin_GroupController extends admin_IndexController
 
 	public function newAction()
 	{
-		if (!empty($_POST)) {
+		if (!empty($_POST) && isset($_POST['groups_name'])
+			&& isset($_POST['groups_description'])) {
 			$data = admin_GroupController::getGroupDataFromPost();
 			$this->save("USVN_Db_Table_Groups", "group", $data);
 		}
@@ -49,7 +51,8 @@ class admin_GroupController extends admin_IndexController
 
 	public function editAction()
 	{
-		if (!empty($_POST)) {
+		if (!empty($_POST) && isset($_POST['groups_name'])
+			&& isset($_POST['groups_description'])) {
 			$data = admin_GroupController::getGroupDataFromPost();
 			$this->save("USVN_Db_Table_Groups", "group", $data);
 		} else {
@@ -57,13 +60,13 @@ class admin_GroupController extends admin_IndexController
 			$this->_view->group = $groupTable->fetchRow(array('groups_name = ?' => $this->getRequest()->getParam('name')));
 		}
 		$this->_render('form.html');
-		$this->_render('../user/attachment.html');
-		$this->_render('../project/attachment.html');
+		$this->_forward('attachUsers');
 	}
 
 	public function deleteAction()
 	{
-		if (!empty($_POST)) {
+		if (!empty($_POST) && isset($_POST['groups_name'])
+			&& isset($_POST['groups_description'])) {
 			$data = admin_GroupController::getGroupDataFromPost();
 			$this->delete("USVN_Db_Table_Groups", "group", $data);
 		} else {
@@ -71,7 +74,52 @@ class admin_GroupController extends admin_IndexController
 			$this->_view->group = $groupTable->fetchRow(array('groups_name = ?' => $this->getRequest()->getParam('name')));
 		}
 		$this->_render('form.html');
+	}
+
+	public function attachUsersAction()
+	{
+		if (!empty($_POST) && $_POST['users_id']) {
+			$groupTable = new USVN_Db_Table_Groups();
+			$group = $groupTable->fetchRow(array('groups_id = ?' => $_POST['group_id']));
+			$group->deleteAllUsers();
+			foreach ($_POST['users_id'] as $user_id) {
+				$group->addUser($user_id);
+			}
+		}
+		$groupTable = new USVN_Db_Table_Groups();
+		$group = $groupTable->fetchRow(array('groups_name = ?' => $this->getRequest()->getParam('name')));
+		$this->_view->group_id = $group->id;
+
+		$groupRow = $group->findManyToManyRowset('USVN_Db_Table_Users', 'USVN_Db_Table_UsersToGroups');
+		$this->_view->attachedUsers = $groupRow->toArray();
+
+		$usersTable = new USVN_Db_Table_Users();
+		$where = $usersTable->getAdapter()->quoteInto('users_login NOT IN (?)', $usersTable->exceptedEntries['users_login']);
+		$this->_view->allUsers = $usersTable->fetchAll($where, "users_login");
 		$this->_render('../user/attachment.html');
+		$this->_forward('attachProjects');
+	}
+
+	public function attachProjectsAction()
+	{
+		if (!empty($_POST) && $_POST['projects_id']) {
+			$groupTable = new USVN_Db_Table_Groups();
+			$group = $groupTable->fetchRow(array('groups_id = ?' => $_POST['group_id']));
+			//$group->deleteAllProjects();
+			foreach ($_POST['projects_id'] as $project_id) {
+				$group->addProject($project_id);
+			}
+		}
+		$groupTable = new USVN_Db_Table_Groups();
+		$group = $groupTable->fetchRow(array('groups_name = ?' => $this->getRequest()->getParam('name')));
+		$this->_view->group_id = $group->id;
+
+		$projectRow = $group->findManyToManyRowset('USVN_Db_Table_Projects', 'USVN_Db_Table_Workgroups');
+		$this->_view->attachedProjects = $projectRow->toArray();
+
+		$projectsTable = new USVN_Db_Table_Projects();
+		$where = $projectsTable->getAdapter()->quoteInto('projects_name NOT IN (?)', $projectsTable->exceptedEntries['projects_name']);
+		$this->_view->allProjects = $projectsTable->fetchAll($where, "projects_name");
 		$this->_render('../project/attachment.html');
 	}
